test: JobDetailSummary test 분리
