+++ front-end/src/domains/JobDetail/JobDetail.test.tsx
@@ -1,4 +1,4 @@
-import { screen, render, waitForElementToBeRemoved } from "@testing-library/react";
+import { screen, render } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import JobDetail from "./JobDetail";
 import { JobResponseMock } from "../../__fixtures__";
@@ -15,91 +15,12 @@ jest.mock("./useJobDetail", () => ({
 }));
 
 describe("JobDetail", () => {
-  const setup = async () => {
-    render(<JobDetail labId={1} />);
-
-    const jobNameHeading = await screen.findByRole("heading", { level: 4, name: /job 이름/i });
-    const jobName = screen.getByText(JobResponseMock.name);
-    const jobStatusHeading = screen.getByRole("heading", { level: 4, name: /Job 상태/i });
-    const jobStatus = screen.getByText(JobResponseMock.status);
-    const jobOwnerHeading = screen.getByRole("heading", { level: 4, name: /Job 등록자/i });
-    const jobOwner = screen.getByText(JobResponseMock.memberName);
-    const assignedServerHeading = screen.getByRole("heading", { level: 4, name: /할당된 서버/i });
-    const assignedServer = screen.getByText(JobResponseMock.gpuServerName);
-    const logHeading = screen.getByRole("heading", { level: 3, name: /log/i });
-
-    return {
-      jobNameHeading,
-      jobName,
-      jobStatusHeading,
-      jobStatus,
-      jobOwnerHeading,
-      jobOwner,
-      assignedServerHeading,
-      assignedServer,
-      logHeading,
-    };
-  };
-
   const mock = ({ detail = JobResponseMock, status = "succeed" } = {}) => {
     mockUseJobDetail.mockReturnValue({
       detail,
       status,
     });
   };
-
-  test("Job이름이 보여진다", async () => {
-    mock();
-    const { jobNameHeading, jobName } = await setup();
-
-    expect(jobNameHeading).toBeInTheDocument();
-    expect(jobName).toBeInTheDocument();
-
-    await waitForElementToBeRemoved(() => screen.getByText(/로그 데이터를 불러오는 중입니다/));
-  });
-
-  test("Job 상태가 보여진다", async () => {
-    mock();
-
-    const { jobStatusHeading, jobStatus } = await setup();
-
-    expect(jobStatusHeading).toBeInTheDocument();
-    expect(jobStatus).toBeInTheDocument();
-
-    await waitForElementToBeRemoved(() => screen.getByText(/로그 데이터를 불러오는 중입니다/));
-  });
-
-  test("Job 등록자가 보여진다", async () => {
-    mock();
-
-    const { jobOwnerHeading, jobOwner } = await setup();
-
-    expect(jobOwnerHeading).toBeInTheDocument();
-    expect(jobOwner).toBeInTheDocument();
-
-    await waitForElementToBeRemoved(() => screen.getByText(/로그 데이터를 불러오는 중입니다/));
-  });
-
-  test("할당된 서버가 보여진다", async () => {
-    mock();
-
-    const { assignedServerHeading, assignedServer } = await setup();
-
-    expect(assignedServerHeading).toBeInTheDocument();
-    expect(assignedServer).toBeInTheDocument();
-
-    await waitForElementToBeRemoved(() => screen.getByText(/로그 데이터를 불러오는 중입니다/));
-  });
-
-  test("Log가 보여진다", async () => {
-    mock();
-
-    const { logHeading } = await setup();
-
-    expect(logHeading).toBeInTheDocument();
-
-    await waitForElementToBeRemoved(() => screen.getByText(/로그 데이터를 불러오는 중입니다/));
-  });
 
   test("status가 loading 일 때 Loading 스피너가 표시된다", async () => {
     mock({ status: "loading" });
